Kalkulacie verzia 1. Pofixovane CRUDy v users, roles, permissions. BaseRepository funkcia getData() ktora pripravuje data pre SimpleTableComponentu.

+++ database/migrations/2020_08_18_161532_create_property_table.php
@@ -16,7 +16,7 @@ class CreatePropertyTable extends Migration
         Schema::create('property', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('amount');
+            $table->float('amount', 10,2);
             $table->text('description')->nullable();
             $table->unsignedBigInteger('claim_id');
             $table->unsignedBigInteger('currency_id');
